Split test command validation fixtures
